+++ world/ViewPlane.cpp
@@ -5,11 +5,11 @@
  *
  */
 ViewPlane::ViewPlane() {
-        top_left     = Point3D(-320.0f, 240.0f, 0.0f);
-        bottom_right = Point3D(top_left.x + hres, top_left.y + vres, 0.0f);
-        normal       = Vector3D(0.0, 0.0, -1.0);
-        hres         = 640;
-        vres         = 480;
+    top_left     = Point3D(-320.f, 240.f, 0.f);
+    bottom_right = Point3D(top_left.x + hres, top_left.y + vres, 0.f);
+    normal       = Vector3D(0.0, 0.0, -1.0);
+    hres         = 640UL;
+    vres         = 480UL;
 }
 
 /**
@@ -17,25 +17,25 @@ ViewPlane::ViewPlane() {
  *
  * @return int
  */
-int ViewPlane::get_hres() const { return hres; }
+size_t ViewPlane::get_hres() const { return hres; }
 
 /**
  * @brief Set the horizontal resolution
  *
  * @param hresVal
  */
-void ViewPlane::set_hres(int hresVal) { hres = hresVal; }
+void ViewPlane::set_hres(size_t hresVal) { hres = hresVal; }
 
 /**
  * @brief Get the vertical resolution
  *
  * @return int
  */
-int ViewPlane::get_vres() const { return vres; }
+size_t ViewPlane::get_vres() const { return vres; }
 
 /**
  * @brief Set the vertical resolution
  *
  * @param vresVal
  */
-void ViewPlane::set_vres(int vresVal) { vres = vresVal; }
+void ViewPlane::set_vres(size_t vresVal) { vres = vresVal; }
